+++ CHANGELOG.md
@@ -25,6 +25,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 - Fixed the CLI writing unneeded logs from `npm` / `pnpm` / `yarn` to the `.log` file.
 - Fixed the app crashing (unhandled error) upon joining two untraceable paths.
 - Fixed the app crashing (unhandled error) upon calling `manager list` and having untraceable paths saved.
+- Fixed `clean` writing twice to the `stdout` what cleanup commands would do.
 
 ## [1.4.2] 27-10-2024